+++ src/components/VPagination/VPagination.ts
@@ -2,3 +2,5 @@ export type VPaginationEvent = {
   itemsPerPage: number;
   page: number;
 };
+
+export type VPagination = {};